Game update. Patch 2.0
Patch 2.0 includes several updates related to the world map, new
components for the enemy, animations, sounds and code improvements.

Added new scripts: ball movements, enemy health bar, teleport to
a specific point (object coordinates) when void OnTriggerEnter2D().

Changed enemy's (pirate) animator for walking, running, idle, attack,
the enemy's controller is written for these animations. Added a new
enemy - chicken that does not interact with the game world and has an
active idle animation. There are 2 chickens on the map.

Remaining changes: scripts are changed to be components of various
objects, increased camera boundaries to see underground chickens,
increased the jump power of the main character for the dynamiс of the
game from 15 to 30.